feat: Write UIModuleInitializerGenerator as incremental generator
