+++ src/Uno.Extensions.Reactive.Generator/UIModuleInitializerGenerator.cs
@@ -1,11 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Text;
-using Microsoft.CodeAnalysis;
+﻿using Microsoft.CodeAnalysis;
 using Uno.Extensions.Generators;
-using Uno.Extensions.Reactive.Config;
-using Uno.Extensions.Reactive.Generator.Dispatching;
 
 namespace Uno.Extensions.Reactive.Generator;
 
@@ -13,13 +7,13 @@ namespace Uno.Extensions.Reactive.Generator;
 /// A generator that generates UI module initialization for the reactive framework.
 /// </summary>
 [Generator]
-public class UIModuleInitializerGenerator : ISourceGenerator
+public sealed class UIModuleInitializerGenerator : IIncrementalGenerator, ICodeGenTool
 {
 	/// <inheritdoc />
-	public void Initialize(GeneratorInitializationContext context) { }
+	public string Version => "1";
 
 	/// <inheritdoc />
-	public void Execute(GeneratorExecutionContext context)
+	public void Initialize(IncrementalGeneratorInitializationContext context)
 	{
 #if DEBUGGING_GENERATOR
 		var process = Process.GetCurrentProcess().ProcessName;
@@ -30,12 +24,38 @@ public class UIModuleInitializerGenerator : ISourceGenerator
 		}
 #endif
 
-		if (GenerationContext.TryGet<UIModuleInitializerGenerationContext>(context, out _) is { } dispatcherContext)
+		var assemblyNameProvider = context.CompilationProvider.Select((compilation, _) => compilation.Assembly.Name);
+		var hasReactiveUIModuleInitializer = context.CompilationProvider.Select((compilation, _) => compilation.GetTypeByMetadataName("Uno.Extensions.Reactive.UI.ModuleInitializer") is not null);
+		context.RegisterSourceOutput(assemblyNameProvider.Combine(hasReactiveUIModuleInitializer), (context, source) =>
 		{
-			foreach (var generated in new UIModuleInitializerGenerationTool(dispatcherContext).Generate())
+			if (source.Right)
 			{
-				context.AddSource(PathHelper.SanitizeFileName(generated.Name) + ".g.cs", generated.Code);
+				var assembly = source.Left;
+				context.AddSource(PathHelper.SanitizeFileName($"{assembly}.ReactiveUIModuleInitializer.g.cs"), $@"{this.GetFileHeader(4)}
+
+namespace {assembly}
+{{
+	/// <summary>
+	/// Initialize provider of dispatcher.
+	/// </summary>
+	/// <remarks>This class ensures that dispatcher has been initialized even if the Reactive.UI package has not been loaded yet.</remarks>
+	[global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Never)]
+	{this.GetCodeGenAttribute()}
+	internal static class __ReactiveUIModuleInitializer
+	{{
+		/// <summary>
+		/// Register the <seealso cref=""DispatcherQueueProvider""/> as provider of <see cref=""IDispatcher""/> for the reactive platform.
+		/// </summary>
+		/// <remarks>This method is flagged with ModuleInitializer attribute and should not be used by application.</remarks>
+		[global::System.Runtime.CompilerServices.ModuleInitializer]
+		public static void Initialize()
+		{{
+			global::Uno.Extensions.Reactive.UI.ModuleInitializer.Initialize();
+		}}
+	}}
+}}
+".Align(0));
 			}
-		}
+		});
 	}
 }
